+++ src/locale/multiple-select-zh-TW.js
@@ -1,5 +1,5 @@
 /**
- * Multiple Select English translation
+ * Multiple Select zh-TW translation
  * Author: Zhixin Wen<wenzhixin2010@gmail.com>
  */